+++ src/globals.cpp
@@ -2,6 +2,6 @@
 // Created by benjamin on 9/26/18.
 //
 
-//#include "globals.h"
-//
-//Sketch SolveSpace::SK = {};
+#include "globals.h"
+
+Sketch SolveSpace::SK = {};
