+++ src/globals.h
@@ -1,15 +1,17 @@
 //
 // Created by benjamin on 9/26/18.
 //
-//
-//#ifndef SOLVESPACE_GLOBALS_H
-//#define SOLVESPACE_GLOBALS_H
-//
-//namespace SolveSpace{
-//
-//extern SolveSpaceUI SS;
-//extern Sketch SK;
-//
-//}
-//
-//#endif //SOLVESPACE_GLOBALS_H
+
+#ifndef SOLVESPACE_GLOBALS_H
+#define SOLVESPACE_GLOBALS_H
+
+#include "solvespaceui.h"
+
+namespace SolveSpace{
+
+extern SolveSpaceUI SS;
+extern Sketch SK;
+
+}
+
+#endif //SOLVESPACE_GLOBALS_H
